+++ Assets/Scripts/Core/GameManagers.cs
@@ -1,25 +1,44 @@
-using Newtonsoft.Json;
+using System;
 using UnityEngine;
 
 public class GameManager : MonoBehaviour
 {
     public static GameManager instance;
 
-    [Header("Current Level informations")]
+    [Header("Levels informations")]
     public LevelData[] alllevelData;
     public LevelData currentLevelData = null;
+
+    [Header("Story")]
     public string nextStory = "";
 
+    [Header("Levels")]
+    public SpawnAlgoData spawnAlgoData;
+    public int nbrSubLevelTotal = 0;
+    public int nbrSubLevelRemaining = 0;
+
+    // strategy
+    public Strategy currentStrategy;
+
+    // deck
+    public string currentDeck;
+    public int currentManaCost;
+
+
+    [Header("Game")]
     public bool isPaused = false;
+
+    [Header("Player")]
+    public int maxLevelReach;
+    public int maxLevelFinish;
 
     private void Awake()
     {
         if (instance == null) {
             instance = this;
             DontDestroyOnLoad(gameObject);
-        } else if (instance != this) {
+        } else if (instance != this)
             Destroy(gameObject);
-        }
     }
 
     // ---- PAUSE / RESUME ----
@@ -40,8 +59,62 @@ public class GameManager : MonoBehaviour
 
     public void SaveLevelConfig(LevelData levelData)
     {
-        BugTracker.Info("Save current level data, current level: '" + levelData.currentlevel + "'.");
+        BugTracker.Info("Save current level data, current level: " + levelData.currentlevel + ".");
         currentLevelData = levelData;
+        SetStrategy();
+
         nextStory = levelData.chaptersBeforeGame;
+    }
+
+    public void SetNextLevel()
+    {
+        if (nbrSubLevelRemaining >= 1) { // sublevel
+            SetNextStrategy();
+            LoadingScene.Instance.ChangeScene("Game");
+            return;
+        }
+
+        // next level
+        int index = Array.IndexOf(alllevelData, currentLevelData);
+
+        maxLevelFinish = currentLevelData.currentlevel;
+
+        index+= 1;
+        if (index >= alllevelData.Length) {
+            BugTracker.Info("No level available, end of the game.");
+            LoadingScene.Instance.ChangeScene("Menu");
+            return;
+        }
+        currentLevelData = alllevelData[index];
+
+        BugTracker.Info("Loading level: " + currentLevelData.currentlevel + ".");
+        SetStrategy();
+        LoadingScene.Instance.ChangeScene("Game");
+    }
+
+    private void SetStrategy()
+    {
+        spawnAlgoData = currentLevelData.spawnAlgo;
+
+        currentStrategy = spawnAlgoData.spawnStartegy[0];
+        currentDeck = spawnAlgoData.deck[0];
+        currentManaCost = spawnAlgoData.manaCost[0];
+
+        nbrSubLevelTotal = spawnAlgoData.spawnStartegy.Length - 1;
+
+        nbrSubLevelRemaining = nbrSubLevelTotal;
+        BugTracker.Info("Number total of sub levels: " + nbrSubLevelTotal + ".");
+        BugTracker.Info("Number remaining of sub levels: " + nbrSubLevelRemaining + ".");
+    }
+
+    private void SetNextStrategy()
+    {
+        nbrSubLevelRemaining -= 1;
+
+        currentStrategy = spawnAlgoData.spawnStartegy[nbrSubLevelTotal - nbrSubLevelRemaining];
+        currentDeck = spawnAlgoData.deck[nbrSubLevelTotal - nbrSubLevelRemaining];
+        currentManaCost = spawnAlgoData.manaCost[nbrSubLevelTotal - nbrSubLevelRemaining];
+
+        BugTracker.Info("Loading next sub levels, remaining: " + (nbrSubLevelRemaining - 1) + ".");
     }
 }