+++ Assets/Scripts/SceneSpecific/Game/Story/StoryData.cs
@@ -1,7 +1,6 @@
-using System.Data;
-using UnityEngine;
+using System;
 
-[System.Serializable]
+[Serializable]
 public struct DialogueLine
 {
     public string backgroundName;
@@ -11,30 +10,8 @@ public struct DialogueLine
     public string text;
 }
 
-[System.Serializable]
+[Serializable]
 public class DialoguesWrapper
 {
     public DialogueLine[] dialogues;
-}
-
-
-[System.Serializable]
-public class LevelData
-{
-    public int currentlevel;
-    public string chaptersBeforeGame;
-    public string chaptersAfterGame;
-    public int nbrSublevel;
-    public bool isBoss;
-
-    // Implement later
-
-    // public MapType mapType;
-    // public DeckData enemyDeck;
-}
-
-[System.Serializable]
-public class LevelsWrapper
-{
-    public LevelData[] levels;
 }